feat(lexer-parser): extract lexer and parser from interpreter issue #223
Move Token, SourceSpan, lex, parse_string_literal, parse_glob_literal
into src/genia/lexer.py and Parser, PRECEDENCE, RESERVED_LITERAL_IDENTIFIERS
into src/genia/parser.py. interpreter.py re-exports all moved names for
backward compatibility. Zero behavior change; 1736 tests pass.
